chore(storage): split manual_compaction_picker to independent file
